+++ src/aui-progressbar/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2058](https://issues.liferay.com/browse/AUI-2058) AUI Progress Bar should use percentage instead of px when horizontal to make it responsive
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)